src/app/components/images/images.component.ts
import { Component, OnInit } from '@angular/core';
import { ImagesService } from 'src/app/services/images.service';
import { Image } from 'src/app/models/image.interface';

@Component({
  selector: 'app-images',
  templateUrl: './images.component.html',
  styleUrls: ['./images.component.sass'],
})
export class ImagesComponent implements OnInit {
  images: Image[] = [];

  constructor(private imagesService: ImagesService) {}

  ngOnInit(): void {
    this.imagesService.getAllImages().subscribe((images) => {
      this.images = images;
    });
  }
}
